pyproject.toml
[project]
name = "local-voice-agent"
version = "0.1.0"
description = "A fully local, privacy-first voice AI assistant with swappable backends"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastrtc[stt]>=0.0.19",
    "kokoro-onnx>=0.4.7",
    "loguru>=0.7.3",
    "ollama>=0.4.7",
    "pyyaml>=6.0",
    "typer>=0.9.0",
]

[project.optional-dependencies]
deepseek = [
    "openai>=1.0",
]
anthropic = [
    "anthropic>=0.18",
]
pyttsx3 = [
    "pyttsx3>=2.90",
    "numpy>=1.24",
]
realtimetts = [
    "RealtimeTTS>=0.3",
    "numpy>=1.24",
]
dev = [
    "pytest>=8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
